Move validateDonationAddress to DisputeValidation
